Factor out handle_input from qrexec-agent-data.c
Changes:
- introduce constants for return codes
- extract an option for "stdio_socket_requested < 2"
